+++ test/test_helper.rb
@@ -13,7 +13,12 @@ Rails.backtrace_cleaner.remove_silencers!
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].each { |f| require f }
 
 def setup_test_fixture
-  @user = User.new(email: 'steve@example.com', password: 'secret', comments: 'my comment')
+  user_klass = Class.new(User)
+  def user_klass.model_name
+    ActiveModel::Name.new(User)
+  end
+
+  @user = user_klass.new(email: 'steve@example.com', password: 'secret', comments: 'my comment')
   @builder = BootstrapForm::FormBuilder.new(:user, @user, self, {})
   @horizontal_builder = BootstrapForm::FormBuilder.new(:user, @user, self, { layout: :horizontal, label_col: "col-sm-2", control_col: "col-sm-10" })
   I18n.backend.store_translations(:en, {activerecord: {help: {user: {password: "A good password should be at least six characters long"}}}})