+++ lib/bootstrap_form/form_builder.rb
@@ -66,9 +66,17 @@ module BootstrapForm
 
     alias_method_chain :file_field, :bootstrap
 
-    def select_with_bootstrap(method, choices, options = {}, html_options = {})
-      form_group_builder(method, options, html_options) do
-        select_without_bootstrap(method, choices, options, html_options)
+    if Gem::Version.new(Rails::VERSION::STRING) >= Gem::Version.new("4.1.0")
+      def select_with_bootstrap(method, choices = nil, options = {}, html_options = {}, &block)
+        form_group_builder(method, options, html_options) do
+          select_without_bootstrap(method, choices, options, html_options, &block)
+        end
+      end
+    else
+      def select_with_bootstrap(method, choices, options = {}, html_options = {})
+        form_group_builder(method, options, html_options) do
+          select_without_bootstrap(method, choices, options, html_options)
+        end
       end
     end