+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Bugfixes:
 Features:
 
   - Allow primary button classes to be overridden (#183, @tanimichi)
+  - Allow to pass a block to select helper in Rails >= 4.1.0 (#229, @doits)
 
 ## 2.3.0 (2015-02-17)